fix(gui): expand grid ranges while building pipeline
