+++ gui/src/client/utils.ts
@@ -1,11 +1,10 @@
 import { doesExist } from '@apextoaster/js-utils';
 import { HighresParams, ModelParams, Txt2ImgParams, UpscaleParams } from '../types/params.js';
-import { ChainPipeline, ChainStageParams } from '../types/chain.js';
+import { ChainPipeline, ChainStageParams, STRING_PARAMETERS } from '../types/chain.js';
 
 export interface PipelineVariable {
   parameter: 'prompt' | 'cfg' | 'seed' | 'steps' | 'eta' | 'scheduler' | 'token';
-  input: string;
-  values: Array<number | string>;
+  value: string;
 }
 
 export interface PipelineGrid {
@@ -13,6 +12,12 @@ export interface PipelineGrid {
   columns: PipelineVariable;
   rows: PipelineVariable;
 }
+
+export const EXPR_STRICT_NUMBER = /^-?\d+$/;
+export const EXPR_NUMBER_RANGE = /^(-?\d+)-(-?\d+)$/;
+
+export const MAX_SEED_SIZE = 32;
+export const MAX_SEED = (2**MAX_SEED_SIZE) - 1;
 
 export function replacePromptTokens(grid: PipelineGrid, params: Txt2ImgParams, columnValue: string | number, rowValue: string | number): {prompt: string} {
   const result = {
@@ -39,9 +44,6 @@ export function replacePromptTokens(grid: PipelineGrid, params: Txt2ImgParams, c
   return result;
 }
 
-export const MAX_SEED_SIZE = 32;
-export const MAX_SEED = (2**MAX_SEED_SIZE) - 1;
-
 export function newSeed(): number {
   return Math.floor(Math.random() * MAX_SEED);
 }
@@ -61,6 +63,37 @@ export function replaceRandomSeeds(key: string, values: Array<number | string>):
   });
 }
 
+export function rangeSplit(parameter: string, value: string): Array<number | string> {
+  const csv = value.split(',').map((it) => it.trim());
+
+  if (STRING_PARAMETERS.includes(parameter)) {
+    return csv;
+  }
+
+  return csv.flatMap((it) => expandRanges(it));
+}
+
+export function expandRanges(range: string): Array<string | number> {
+  if (EXPR_STRICT_NUMBER.test(range)) {
+    // entirely numeric, return after parsing
+    const val = parseInt(range, 10);
+    return [val];
+  }
+
+  if (EXPR_NUMBER_RANGE.test(range)) {
+    const match = EXPR_NUMBER_RANGE.exec(range);
+    if (doesExist(match)) {
+      const [_full, startStr, endStr] = Array.from(match);
+      const start = parseInt(startStr, 10);
+      const end = parseInt(endStr, 10);
+
+      return new Array(end - start).fill(0).map((_value, idx) => idx + start);
+    }
+  }
+
+  return [];
+}
+
 // eslint-disable-next-line max-params
 export function makeTxt2ImgGridPipeline(grid: PipelineGrid, model: ModelParams, params: Txt2ImgParams, upscale?: UpscaleParams, highres?: HighresParams): ChainPipeline {
   const pipeline: ChainPipeline = {
@@ -75,8 +108,8 @@ export function makeTxt2ImgGridPipeline(grid: PipelineGrid, model: ModelParams, 
     tiles: 8192,
   };
 
-  const rows = replaceRandomSeeds(grid.rows.parameter, grid.rows.values);
-  const columns = replaceRandomSeeds(grid.columns.parameter, grid.columns.values);
+  const rows = replaceRandomSeeds(grid.rows.parameter, rangeSplit(grid.rows.parameter, grid.rows.value));
+  const columns = replaceRandomSeeds(grid.columns.parameter, rangeSplit(grid.columns.parameter, grid.columns.value));
 
   let i = 0;
 
@@ -108,8 +141,8 @@ export function makeTxt2ImgGridPipeline(grid: PipelineGrid, model: ModelParams, 
       ...params,
       ...model,
       ...tiles,
-      height: grid.rows.values.length,
-      width: grid.columns.values.length,
+      height: rows.length,
+      width: columns.length,
     },
   });
